Added symbol table
The symbol table is a hash table, where collation area is implemented like a
double linked list with a dummy header. All the item presents in the table
are stored in a single linked list, so it is easy run over the last symbols
and remove all of them each time a block ends.