Refactoring of specifier function
The code was really ugly and it was impossible use the type logic in other
part (for example in type name rules). In first step we split the type
qualifiers and modifiers in a different struct, and let the base type as a
tree, which is generated using only the 3 type compound operators: [], ()
and *.

So we will have 2 different structs now, ctype which holds the
modifiers (and it will form part of the symbol) and btype, which will be
built using the type operators.