Removed pointers from types.
Original idea was having an unique tree where all the types were built. But
this idea is not possible due to the pointer modifiers. The modifiers can
appears in any order, so a "int *cont volatile" is the same that a
"int *volatile const", but using the actual tree system, they will be
represented using different pointers, so the system is broken. After this
patch the system is a continuos memory leak, but this is not important in
this moment.