Removed type struct
The ctype struct supplant the place of the old struct type, because ctype
has the information about the type qualifiers needed by pointers. This imply
that different variables will share the same ctype struct (pointers will
share the same base type), so it is necessary a reference count system. The
reference count system is implement using the new functions newctype and
delctype.