Rewriting of symbol module
Before of this re-writing the module was too much complex and has some
errors. This module is clever now and doesn't break in any
conditions. Instead of having different hash for each name space, we will
have only one (splitting the hash using the ns field), and instead of having
and array marking the end of each context, we have a field in the symbol
which indicate the context of each one.