Fixed bug reading pointer modifiers
The code was reading only one character in each iteration of the loop, and
this was working in the case of indirection without modifiers (ex: int *v),
but was erroneus in the case of two indirection levels (ex: int *const
v). It is necessary one next() call for each '*' token and other call for
each 'const', 'volatile' or 'restrict'.