Added support for different namespace in the symbol table
The symbol table hold symbols relative to different namespaces, (keywords,
identifiers, labels ...). So we need can lookup only in one namespace. When
the lexical analysis is done, we don't have this information, but we need
the symbol pointer requested later. The aproach used in this patch consist
in assume number of duplicated symbol names is low, so we usually will catch
the correct one without be worried in the namespace, and later, in the moment
that we have enough information of the symbol, check is the namespace is
correct or we have to do a new lookup with the correct namespace.