Rewrite of ahead function
ahead is the function used for read ahead a token in order to can difference
the rule to apply in some cases (for example in the case of and IDEN which
can be a label or a expression). This imply have to read the full token, and
maybe save the content of yytex, for later restauring, and some ugly stricks
in next() function (it has to test it there is some previous token due to
ahead()).

Instead of doing the test of the token we can test only the next character
(after skipping blanks of course), because this will be enough (for example
in the label case, next character must be a ':'), and we don't have to worry
about yytext or near calls to next() (character read is returned to the
stream using ungetc).