Added control about nesting levels
If we don't want have a stack colition with the heap, we should check the
nesting level, because in other case the recursion will be controlled by the
user, and this is dangerous. It is also important not only check the deep
level, and write why we nest (while, do, for, switch), because this
information will be used later for break, continue, case and default.