Remove context parameter from lookup
We will allways look symbols based in the name and maybe in the
namespace, we don't need the context in the lookup function.
The unique case where we add a symbol in a different context of the
actual, is in the label case (labels are function scope symbols), so we
need a insert function which allow us modify the context of a symbol.