Fix del_ctx function
del_ctx must decrement the context number after deleting the symbols
of the context, because in other case the condition of the loop is not
working fine. We can not delete the symbols in del_ctx because they are
used in the parsing tree, but we must free them in some moment, so
after printing the tree is a good moment.