Move curfun to flow.c
curfun is a useful variable which allow us to can access to the
current function and could check the return type of the function. This
information is only needed in flow.c, and it is the best place for
knowing when we are defining a new function (due to function()).