Add option for signed characters
ansi c says that is implementation defined the signess of characters,
and since we have addede a c_sign bit in ctype, it is not a bad idea
adding a flag which allows to the user defines the signess of characters.
This patch also fixes the problem of filling c_sign in other integers
cases.