Add namespace function
These function returns to the user the symbol in the desired
namespace. Like the first lookup is always done in the lexical
part, in that moment we don't have the information for selecting
the correct namespace, so it selects the last defined symbol, and
not always it is the symbol we are looking for.

This function check if the symbol is in the correct namespace,
and if it is necessary it installs it.