Allow registration of defined symbols in namespace
namespace registers the symbol detected in the lexical analysis
in a namespace. At the moment, it raised an error when we try
register again a name in the same namespace, but this is not
the behaviour we want in the case of forward declarations, so
we need a way of indicating this case. After this change
alloc > 0 means allocate a new symbol if no defined, alloc == 0
means no allocate a new symbol if no defined and raises an error
and alloc < 0 means no raises an error when the symbol is already
defined.