Add struct and unions declarators
These declarators are very diferents of the previous due to
they define a new namespace, so the thecniques used previously
cannot be applied here. We have to define a new namespace with
each new struct declared. We defined the first struct namespace
as NS_ANY+1, so we will not have any collision with the anothers
namespaces.