Simplify prtree
Due to the new representation of the tree, it is easiar run over
a tree, so we can remove the nchild field in the representation
table.

It is also no needed the prtree_helper function, because each increment
of indent is always followed of a decrement, so we can use only a function
because when prtree returns indent holds the same value it had in the
beginning.