Add octal and hexadecimal numbers
C support numbers in decimal, octal and hexadecimal. Octal numbers
begin with 0, and hexadecimal numbers begin with 0x. This patch
adds support for octal and hexadecimal, although it doesn't check
the base of the number when insert in the symbol table, so maybe
two numbers with different bases are allocated to the same symbol.