Remove insert
insert does the same than lookup, but it never installs a new symbol,
so we have duplicated the code in both functions. insert is used only
once in all the code, so add a new parameter doesn't seems me the
correct way, because it implies modify a lot of code, and it is possible
also get some performance penalty. The solution used is insert in the
namespace parameter of lookup a new parameter which indicatens if it is
needed install the symbol, using a negative number when it is not needed.