Remove reference count in types
In the types representation we were using a reference counting system,
but difference instances of the same object (different memory positions)
could exits, so having the reference system was paintful, and the
utility were reduced because types were shared only between variables
defined in the same list declaration, that is not the usual case. This
patch removes the reference count system and  always duplicate the types
in declaration lists.