Some changes in function names
We were calling struct_spec to the function called from spec, but we
were calling union_dcl to the similar function for unions, and it was
confusing, because struct_dcl was the function where struct fields
were defined. This patch change struct_dcl to field_dcl, and struct_spec
to struct_dcl, so the code is more clever now.