Use a different hash for keywords
With the previous implementation it was necessary look in the
symbol table in order to know if a identifier was a keyword
or not. This search generated the problem of not know yet
the correct namespace, and due to this problem it was necessary
the ugly code of namespace.
This patch is the first step to remove this shit, and uses
a new hash only for the keywords, so the problem of the
namespaces will be fixed because we push the search to other
moment when we already have the needed information.