Remove namespace function
These function was a big mess, and it was used only because
having the keywords in the symbol table caused a lot of problems.
This new version allows don't use yyval anymore, which is good
because it was generating some dependencies problems.
Due to these changes label definition was also affected, so
newlabel is no longer needed because lookup does perfectly
this job.