Use a different namespace for each struct/union
The code had a bug and it was using the same namespace for all
the fields of all the structs, so it was causing problems. This
patch fixes this problem, and also remove the non used structbuf
and structp (the value of the namespace is saved in the stack
due to the recursion).