Pass type struct to specifier
This is the first step for removing the storage specifier
from the type struct. We need this step because we will
have a pointer to the type and a pointer to the storage,
while specifier was a void function returning a pointer
to ctype, and it is no longer valid.

This change implies we can not test against NULL in specifier,
because it is not going to be true anymore, so we introduce two
new fields, that indicate if type or storage are defined. This
fields are going to be splitted later in two different structs.