Simplify listdcl function
We have two different cases in listdcl, variables and functions.
When we have a variable we can find a initializer, but when you have
a function you can have the body of the function. This simplification
allows to have a linear code and a common exit point.

Since this construction is very uncommon, I think isn't a big
problem and we can 'extend' the language in this direction ;).