Fix struct type initialzation
aggregate() receives tp,a pointer to an automatic variable, so this
pointer cannot be used in a type struct, because the new type
will have a longer life than the function were was defined the
value pointed by tp. It is necessary allocate a new type and
assign it later.