Move check of tag already defined out of aggregate
aggregate function is going to be called each time an aggregate declaration
begins, in normal declaration and in forward declarations, but the check
of 'tag already defined' must be done only in the case of normal
declaration. The place where we know if the declaration is normal or
forward is in structdcl after the aggregate() call, so the check is
moved there.