Add namespace parameter to decl()
decl() parses a declaration, that can be an usual declaration of a variable
or a field of a struct of union. The only difference in both cases is
the namespace where symbols are going to be stored, so this is a first
step in order to unify decl() and fielddcl().