Do a prefetch of 1 token
C is a LR(2) language so it means we need 2 tokens in order to
decide what is the next step. We have solving this problem with
backtracking and a tricky ahead, that was used only to detect that
the proximus token was ':' or not. Since we want to remove the
backtracking we need a real ahead() function, and the better
way is having a prefetch token.