Convert pointers to address of arrays in ary()
We have in the stack the address of the variables, so when we
want to access to some variables, the first thing we have to do
is fetch the value of the variable from the address. In the case
of a pointer, we have to fetch the value of the pointer, and
them fetch the content of the pointer.