Don't swap integer float operands
The swap was done due to pointers, because in that
case there are too much cases, and we didn't want to
replicate the effort in cases as PTR + INT or INT + PTR,
but it generates problems in cases as comparisions. This
is the first step, where we remove the swap only for int
and float types.