Change name of ISNODELOG to ISNODECMP
ISNODELOG stand for "is node logical", but it was not a good
mnemonic. ISNODECMP stands for "is node comparision", and it is
a better mnemonic, but is not true, because it is also true for
logical and and logical or. We are going to use ISNODECMP
because is more legible.