Fix compare()
A comparision has always a integer type, but the operation
done in the Node may not be a integer operation. The node
of the type also help in order to know what version of
the operator must be done, so in this case we have to
put the type of the operands, and let that eval() convert
the Node type to integer where it is necessary.