Remove ns field of symbols
This field was added in order to can define a namespace for
every aggregate, but it was a bad way of doing it. It is a good
idea to create a new context only for the declaration of aggregates,
because we don't want to create symbols for them, we only want to
take the symbol name and the type in order to create the fields
of the struct.