Simplify stmt
We can use a function pointer for this function, but then we need
to pass the state to all the possible statements. This is not a
problem, because we can ignore the parameters we are not using,
and the code of stmt will be smaller because we have only one
function call.