Allow tags with same name that a previous typedef
Tags have a separate namespace, so it is correct this combination,
but the problem is the lexer returns a different token for an
identifier after a typedef for it. For example:

	main()
	{
		typedef int pepe;
		struct pepe {
			int i;
			char c;
		} h;
	}

The easier solution is a hack, where we modify the grammar to
accept an IDEN or a TYPE when creating tags.