Add eqtype()
This function compare two types and see if they are the same
type or not. This function is needed because the comparision
of pointers can be difficult, because if they pointed to a
function they are going to point to different structures,
because different functions with the same type have different
Type structure (this is done because is hard to generate a
hash in this case).