Remove decay in convert()
convert() do all the needed operations to convert an expression
of a type in another type. It was generating a decay when the
destination type was an array or an function, but it was incorrect,
because it was generating a decay of the node, that surely was
not an array or a function. The correct here is return NULL,
because we should not receive such types as destination
types.