Forbids castings to array or functions
These casting are explicitily forbidden in the standard, and
it has sense, since we can not change the type of an object
memory.

There is no legal use of a function type name, and it makes
ambigous the gramatic, because the parentheses can be nested
directdcl or a fundcl. So the simpler solution is to forbid
always the type names of functions.