Make explicit checks of type in castings and sizeof
The grammar we are using doesn't allow specify a function type
name, but C standard allows it. This is a desviation from the
standard, which is obvius wrong here. Accepting function types
in typenames (or abstract declarators) make the grammar ambiguous
because it is impossible difference between:

	(int (f))  -> function returning int with one parameter of type f
	(int (f))  -> integer variable f

Function type names are stupid, because they are used as an alias
of the function pointer types, but it makes stupid that something
like sizeof(int (int)) is not allowed (because here should be
understood as the size of a function), but it is allowed f(int (int))
because it is understood as a parameter with function pointer type.

Although we don't allow function types names, we add checks against
function types in sizeof and casting, becaues id adds documentation
to the code and prepare it for the future.