Fix parsing of 'T' elements
'T' is used for static variables that are not global, and they
can appear in outer contex or in the context of a function. If
they appear in the context of function they have to be stored
in the local pool, and if they appear in the outer then they
must be stored in the global spool.