Generate unexpected in primary()
The parser was returning NULL when it was no able of parsing a expression,
and this was used in for statements for empty expressions, but it
generate segmentation faults in some cases. It is better check when
the expression can be NULL because is faster (we don't have to call
the recursive functions of expr()), and it is more secure because we
don't have to check against empty expressions in all the places where it
is called.