Fix error creating types of functions
Function types had a field for the parameters of the function, but
this field was not set. This patch fixes this problem and changes the list
of parameters to an array. We can use a dinamic array here, but the limit
is big enough (if someone uses more parameters then he should go to the
hell).