Fix creation of data type in functions and structures
There were different reasons why this creation was incorrect in both
cases. In the case of structs we were decrementing the number of
elements, when it was correct, and this create very big numbers
in the case of 0 elements. In the case of the functions the problem
was the ternary operator, that is not neeeded at all, because we already
check that the number of elements is bigger than 0.