Simplify node() function
Node() creates a new node, and it was accepting a value that
was the kind of node, and a vararg list of descendant. This created
a lot of problems, and some ugly unions were needed. This patch
simplifies the Node struct and make that node only receives a node
operator.