Add longjmp point of recover
This is the first commit of the non exit feature.
At this point when there is an error the program try
an external declaration again. This is a very bad
error recovery, because it is going to generate
error for all the symbols after the first, but it is
a first step.