Avoid trashing cpp symbols until be sure is correct
With the previous implementation, the status of the u.s field
could be corrupted, because it was freed, and it was not
put to NULL, and the symbol was still defined.
This new solution removes the problem don't looking the
symbol until be sure that the full #define directive
is correct, so it is impossible to have a wrong state.