Fix expansion of macros
When a macro is expanded it is possible (and common) to have
more available characters in the current line, so we cannot
be sure that after a delinput() we have to read a line from
the current line, because it is possible to return from a
macro expansion. This change needs also to discard the
content of the line after processing a preprocessor directive.