Rewrite and simplify the embedded preprocessor
The preprocessor dealt macro expansions in the same way that
includes. This was done in this way as a first attemp to avoid
recursion, but it only avoided direct recursion. Implement the
algorithm described in ANSI C is really complex, and the idea
of the embedded preprocessor was to have a fast and simple
preprocessor, so it is better to remove any idea of trying to
detect recursivity. It will simplify a lot the input/output
functions.
This version of the preprocessor uses the lexer of the compiler,
so the code is not so hardcore like the original which did all
the parsing directly with char pointers.