Deal size_t like any other type
The important point in the types in the current
implementation is the value of the fields, so two
types can be represented by different pointers,
with different values that if they have the same
values then they will be the same type. For this
reason is better keep size_t like a common type.