Rewrite decl.c
decl.c had several function that were very similar between them,
basically the different kind of declarations in a C program:

	- externals
	- internals
	- parameters
	- struct/union fields

This version puts all the logic in only one function, dodcl(),
which receives different parameters to adapt to the different
situations. The idea was taken from kencc and lcc.