Emit function types before emiting functions
Function types are needed when a function is called, because
the ABI for a function like f(...) can be different to the
ABI for f(int), even in the case the receives only one
parameter.

This change emits parameters after emiting the function, so
we can keep the storage specifier of the parameters and free
P for pointers.