Fix fundcl()
Fundcl() has to take the decision when it is needed the parameter
context or not, and the only way is to check if the next token
is a allowed one in a function definition. Ideally the
context should be poped before calling expect(), but we know
that after a parameter list cannot go a identifier, so it is
not a problem.