Fix cast()
Cast() is a problematic function because we need see two tokens
of the input, but the second token can be a typename, so we can
not use ahead() because it only can return the next character
of the input. The solution is to duplicate the rules.
When cast() sees a '(' it means that can be a cast, or can be a
parentheses expression, so it accepts the '(' and calls expr(),
but due to the lost of recursivity the actions of postfix were
not applied, so the solution is to call a modified version of
postfix() before returning.