Rewrite symbol table again
Lookup() was inserting in the table when a symbol was not defined.
It was a bad idea because there are sometimes where the namespace
is not known in the lexer time. It was creating symbols in the
incorrect namespace, and it was difficult to re use these symbols,
and in some times they were masquerading other symbols. The solution
is allocate when a symbol is not found but don't insert it in the
symbol table.