Fix define()
After the last symbol table rewritten, define in cpp was not working,
because the semantic of the table was totally different. Symbols are
not ever declared with lookup(), a explicit call to install is needed.
For the same reason, it is also necesary check in iden() if we are in
cppmode, because in other case we can call to nextsym() in all the
symbols defined in the preprocessor.