Fix nextsym()
Nextsym() is used when a macro with arguments is found without arguments,
and it is needed to look for a definition with less priority. There were
two different errors in nextsym(), because in case of finding the less
priority symbol it was returning the more priority, and in case of not
finding a less priority symbol it was installing a new symbol, but it
should be only allocated.